requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
diskcache==5.6.3
exceptiongroup==1.3.1
fastapi==0.124.4
filelock==3.16.1
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==2.1.5
numpy==1.24.4
packaging==25.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.4
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.44.0
starlette-context==0.3.6
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
